requirements.txt
streamlit
python-docx
docxtpl
pdfkit
# spacy
pdfplumber

